\name{beetle}
\alias{beetle}
\docType{data}
\title{Japanese beetle data}
\description{
Numbers of Japanese beetle larvae per square found in the top foot of soil
of an 18 x 8 foot area of a field planted with maize.  The columns of the matrix
correspond to the direction of cultivation of the field; the maize rows were sown
4 feet apart.
}
\usage{data(beetle)}
\format{
  The format is:
 num [1:18, 1:8] 0 2 3 1 5 3 5 3 2 3 ...
}
\source{Unpublished lecture notes, Imperial College, London}
\examples{
data(beetle)
}
\keyword{datasets}
