\name{births}
\alias{births}
\docType{data}
\title{Birth Times}
\description{
Times spent in delivery suite by 95 women giving birth at the John Radcliffe Hospital,
Oxford.  The data were kindly provided by Ethel Burns.
}
\usage{data(births)}
\format{
  A data frame with 95 observations on the following 2 variables.
  \describe{
    \item{day}{Day on which woman arrived}
    \item{time}{Time (hours) spent on delivery suite}
  }
}

\source{
Davison, A. C. (2003)  \emph{Statistical Models}.  Cambridge University Press. Page 18.
}

\examples{
data(births)
}
\keyword{datasets}
