\name{blood}
\alias{blood}
\docType{data}
\title{Blood Group Data}
\description{
Data on the incidence of blood groups O, A, B, and AB
in 12 studies on people living in Britain or of British
origin living elsewhere.
}
\usage{data(blood)}
\format{
  A data frame with 12 observations on the following 4 variables.
  \describe{
    \item{O}{Number of persons with blood group O}
    \item{A}{Number of persons with blood group A}
    \item{B}{Number of persons with blood group B}
    \item{AB}{Number of persons with blood group AB}
  }
}
\source{Taylor, G. L. and Prior, A. M. (1938)
Blood groups in England.  \emph{Annals of Eugenics}, \bold{8}, 343--355.
}
\keyword{datasets}
