\name{cardiac}
\alias{cardiac}
\docType{data}
\title{Mortality Rates for Cardiac Surgery on Babies at 12 Hospitals}
\description{
The title should be self-explanatory.
}

\usage{data(cardiac)}
\format{
  A data frame with 12 observations on the following 2 variables.
  \describe{
    \item{r}{Number of deaths}
    \item{m}{Number of operations}
  }
}
\source{Spiegelhalter, D. J., Thomas, A., Best, N. G. and W. R. Gilks
(1996)  \emph{BUGS 0.5 Examples Volume 1 (Version ii)}.  
Cambridge: MRC Biostatistics Unit.  Page 15.
}
\references{
Davison, A. C. (2003)  \emph{Statistical Models}.  Cambridge University Press.
Page 579.}
\keyword{datasets}
