\name{fir}
\alias{fir}
\docType{data}
\title{Counts of Balsam-fir Seedlings}
\description{
    The number of balsam-fir seedlings in each quadrant of a grid of
     50 five foot  square quadrants were counted.  The grid consisted
     of 5 rows of 10 quadrants in each row.
}
\usage{data(fir)}
\format{
  A data frame with 50 observations on the following 3 variables.
  \describe{
    \item{count}{The number of seedlings in the quadrant}
    \item{row}{The row number of the quadrant}
    \item{col}{The quadrant number within the row}
  }
}
\source{
Davison, A.C. and Hinkley, D.V. (1997)  \emph{Bootstrap Methods and
     Their Application}. Cambridge University Press.}
\references{
Davison, A. C. (2003)  \emph{Statistical Models}.  Cambridge University Press.
Page 641.}
\keyword{datasets}
