\name{get.alpha}
\alias{get.alpha}
\title{Estimate Alpha from Data}
\description{
Estimate prior value of a parameter from the data.  
}
\usage{
get.alpha(d)
}
\arguments{
  \item{d}{A data frame with vector components \code{y} and \code{x} of the same length.}
}
\value{
  Prior value of a parameter, estimated from the data.
}
\seealso{\code{\link{poi.beta.laplace}}}
\keyword{distribution}
