\name{intron}
\alias{intron}
\docType{data}
\title{Intron Gene Sequence}
\description{
Sequence of 1572 bases from first human preproglucagon gene
}
\usage{data(intron)}
\format{
  This is a factor with 4 levels,  "A","C","G","T"
}
\source{
Avery, P. J. and Henderson, D. A. (1999) Fitting Markov chain models to
  discrete state series such as DNA sequences.
\emph{Applied Statistics}, \bold{48}, 53--61.
}
\references{
Davison, A. C. (2003)  \emph{Statistical Models}.  Cambridge University Press.
Page 226.
}
\examples{
data(intron)
## maybe str(intron) ; plot(intron) ...
}
\keyword{datasets}
