\name{jelinski}
\alias{jelinski}
\docType{data}
\title{Jelinski and Moranda Data on Software Failures}
\description{
These are times in days between sucessive failures of a piece of
software developed as part of a large data system.  The software
was released after the first 31 failures.  The last three failures
occurred after release.
}
\usage{data(jelinski)}
\format{
  The format is:
 num [1:34] 9 12 11 4 7 2 5 8 5 7 ...
}
\source{Jelinski, Z. and Moranda, P. B. (1972)  
Software reliability research.  In W. Freiberger (ed), 
\emph{Statistical Computer Performance Evaluation}.  London: 
Academic Press.  Pages 465--484.
}
\references{
Davison, A. C. (2003)  \emph{Statistical Models}.  Cambridge University Press.
Page 299.}
\keyword{datasets}
