\name{nematode}
\alias{nematode}
\docType{data}
\title{Nematode Data}
\description{
Numbers of nematodes invading individual fly larvae for
various numbers of initial challengers.
}
\usage{data(nematode)}
\format{
  A data frame with 29 observations on the following 3 variables.
  \describe{
    \item{m}{Number of challengers}
    \item{r}{Number of invading nematodes}
    \item{y}{Number of fly larvae}
  }
}
\source{
Faddy, M. J. and Fenlon, J. S. (1999) Stochastic modelling of the invasion
  process of nematodes in fly larvae.
\emph{Applied Statistics}, \bold{48}, 31--37.
}
\references{
Davison, A. C. (2003)  \emph{Statistical Models}.  Cambridge University Press.
Page 295.
}
\examples{
data(nematode)
}
\keyword{datasets}
