\name{pipette}
\alias{pipette}
\docType{data}
\title{Red Blood Cell Data}
\description{
Numbers of red blood cells counted by five doctors using ten sets of apparatus.
}
\usage{data(pipette)}
\format{
  A data frame with 50 observations on the following 3 variables.
  \describe{
    \item{apparatus}{Factor with ten levels}
    \item{doctor}{Factor with five levels}
    \item{y}{Number of red blood cells}
  }
}
\source{
Unpublished lecture notes, Imperial College, London.
}
\references{
Davison, A. C. (2003)  \emph{Statistical Models}.  Cambridge University Press.
Page 462.
}
\keyword{datasets}
