\name{quake}
\alias{quake}
\docType{data}
\title{Japanese Earthquake Data}
\description{
Times and magnitudes (Richter scale) of 483 shallow earthquakes in an
offshore region east of Honshu and south of Hokkaido, for the period
1885--1980.
}
\usage{data(quake)}
\format{
  An irregular time series with earthquake
  \describe{
    \item{time}{in days since start of 1885}
    \item{mag}{magnitude (Richter scale)}
  }
}
\source{
Ogata, Y. (1988) Statistical models for earthquake occurrences and 
residual analysis for point processes.
\emph{Journal of the American Statistical Association}, \bold{83}, 9--27.
}
\references{
Davison, A. C. (2003)  \emph{Statistical Models}.  Cambridge University Press.
Page 289.
}
\keyword{datasets}
