\name{soccer}
\alias{soccer}
\docType{data}
\title{Soccer Scores from English Premier League, 2000-2001 Season}
\description{
These are scores for the 380 fixtures in the English Premier League,
2000--2001.
}
\usage{data(soccer)}
\format{
  A data frame with 380 observations on the following 7 variables.
  \describe{
    \item{month}{Month of match}
    \item{day}{Day of match}
    \item{year}{Year of match}
    \item{team1}{Home team}
    \item{team2}{Away team}
    \item{score1}{Goals scored by home team}
    \item{score2}{Goals scored by away team}
  }
}
\source{
http://www.soccerbase.com/footballlive/
}
\references{
Davison, A. C. (2003)  \emph{Statistical Models}.  Cambridge University Press.
Page 499.}
\keyword{datasets}
